refactor: InstructionsDecoder out of DISASM VM
